refactor(*): move all components for dataset page to `components/dataset` folder
