+++ lib/components/dataset/Body.js
@@ -1,13 +1,13 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Base from './Base'
-import Json from './Json'
-import Spinner from './Spinner'
-import DatasetRefProps from '../propTypes/datasetRefProps'
-import Button from './Button'
-import HandsonTable from './HandsonTable'
-import { extractSchema } from '../selectors/dataset'
+import Base from '../Base'
+import Json from '../Json'
+import Spinner from '../Spinner'
+import DatasetRefProps from '../../propTypes/datasetRefProps'
+import Button from '../Button'
+import HandsonTable from '../HandsonTable'
+import { extractSchema } from '../../selectors/dataset'
 import BodyReadOnly from './BodyReadOnly'
 
 export default class Body extends Base {
