+++ lib/components/dataset/BodyReadOnly.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Base from './Base'
+import Base from '../Base'
 
 export default class BodyReadOnly extends Base {
   template (css) {
