+++ lib/components/dataset/Commits.js
@@ -2,10 +2,10 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { debounce } from 'lodash'
 
-import Base from './Base'
-import List from './List'
-import CommitItemContainer from '../containers/CommitItem'
-import Spinner from './Spinner'
+import Base from '../Base'
+import List from '../List'
+import CommitItemContainer from '../../containers/CommitItem'
+import Spinner from '../Spinner'
 
 export default class Commits extends Base {
   constructor (props) {
