+++ lib/components/dataset/Dataset.js
@@ -3,16 +3,16 @@ import React from 'react'
 import { Route, Switch } from 'react-router'
 import PropTypes from 'prop-types'
 
-import DatasetRefProps from '../propTypes/datasetRefProps'
+import DatasetRefProps from '../../propTypes/datasetRefProps'
 
-import Base from './Base'
+import Base from '../Base'
 
 import DatasetHeader from './DatasetHeader'
 import Render from './Render'
 import Export from './Export'
 import DatasetRouter from './DatasetRouter'
 
-import Spinner from './Spinner'
+import Spinner from '../Spinner'
 
 // DATASET_DEBUG = 1 will print the if statements and functions run in componentWillMount and componentWillRecieveProps
 // DATASET_DEBUG = 2 will print the above and also the props, or the props and nextProps, in the componentWillMount and componentWillRecieveProps
