+++ lib/components/dataset/DatasetHeader.js
@@ -2,14 +2,14 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { withRouter } from 'react-router-dom'
 
-import DatasetRefProps from '../propTypes/datasetRefProps'
-import Button from './Button'
-import DatasetName from './DatasetName'
-import Header from './Header'
+import DatasetRefProps from '../../propTypes/datasetRefProps'
+import Button from '../Button'
+import DatasetName from '../DatasetName'
+import Header from '../Header'
 
-import { defaultPalette } from '../propTypes/palette'
-import { addActiveToLink } from '../utils/links.js'
-import Base from './Base'
+import { defaultPalette } from '../../propTypes/palette'
+import { addActiveToLink } from '../../utils/links.js'
+import Base from '../Base'
 
 class DatasetHeader extends Base {
   constructor (props) {
