+++ lib/components/dataset/DatasetRouter.js
@@ -6,18 +6,18 @@ import PropTypes from 'prop-types'
 
 import DatasetRefProps from '../propTypes/datasetRefProps'
 
-import Base from './Base'
+import Base from '../Base'
 
-import NavLinks from './NavLinks'
+import NavLinks from '../NavLinks'
 import Structure from './Structure'
 import Meta from './Meta'
-import TransformContainer from '../containers/Transform'
-import CommitsContainer from '../containers/Commits'
+import TransformContainer from '../../containers/Transform'
+import CommitsContainer from '../../containers/Commits'
 import Body from './Body'
-import VizContainer from '../containers/Viz'
+import VizContainer from '../../containers/Viz'
 import Overview from './Overview'
 
-import { addActiveToLink } from '../utils/links.js'
+import { addActiveToLink } from '../../utils/links.js'
 
 class DatasetRouter extends Base {
   constructor (props) {
