+++ lib/components/dataset/Export.js
@@ -1,8 +1,8 @@
 import React from 'react'
 
 import PropTypes from 'prop-types'
-import Button from './Button'
-import Base from './Base'
+import Button from '../Button'
+import Base from '../Base'
 
 export default class Export extends Base {
   template (css) {
