+++ lib/components/dataset/Meta.js
@@ -1,9 +1,9 @@
 import React from 'react'
 
 import MetadataViewer from './MetadataViewer'
-import Base from './Base'
-import MetaProps from '../propTypes/metaProps.js'
-import { defaultPalette } from '../propTypes/palette'
+import Base from '../Base'
+import MetaProps from '../../propTypes/metaProps.js'
+import { defaultPalette } from '../../propTypes/palette'
 
 export default class Meta extends Base {
   template (css) {
