+++ lib/components/dataset/MetadataViewer.js
@@ -1,24 +1,34 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import { defaultPalette } from '../propTypes/palette'
+import { defaultPalette } from '../../propTypes/palette'
 
-import Base from './Base'
-import Json from './Json'
+import Base from '../Base'
+import Json from '../Json'
 
-export default class StructureViewer extends Base {
+export default class MetadataViewer extends Base {
   constructor (props) {
     super(props)
 
     this.stdFields = [
-      'checksum',
-      'errCount',
-      'entries',
-      'format',
-      'length'
+      'title',
+      'theme',
+      'keywords',
+      'description',
+      'license',
+      'accessPath',
+      'language',
+      'citations',
+      'contributors',
+      'accrualPeriodicity',
+      'downloadPath',
+      'homePath',
+      'identifier',
+      'readmePath',
+      'version'
     ]
 
-    this.ignoreFields = ['qri', 'path', 'schema'];
+    this.ignoreFields = ['qri', 'path'];
 
     ['renderTable', 'renderValue'].forEach((m) => { this[m] = this[m].bind(this) })
   }
@@ -56,25 +66,22 @@ export default class StructureViewer extends Base {
   }
 
   template (css) {
-    const { structure, stdOnly } = this.props
+    const { metadata, stdOnly } = this.props
 
-    const std = this.stdFields.filter((key) => !!structure[key])
-    const extra = Object.keys(structure).filter((key) => {
+    const std = this.stdFields.filter((key) => !!metadata[key])
+    const extra = Object.keys(metadata).filter((key) => {
       return !(~this.stdFields.findIndex((sKey) => (key === sKey)) || ~this.ignoreFields.findIndex((iKey) => (key === iKey)))
     })
 
     return (
       <div className={css('wrap')}>
-        <h5 className={css('title')}>Structure</h5>
-        {this.renderTable(css, std, structure)}
+        <h5 className={css('title')}>Standard Metadata</h5>
+        {this.renderTable(css, std, metadata)}
 
         {!stdOnly && (extra.length > 0) && <div>
-          <h5 className={css('title')}>Schema</h5>
-          {this.renderTable(css, extra, structure)}
+          <h5 className={css('title')}>Additional Metadata</h5>
+          {this.renderTable(css, extra, metadata)}
         </div>}
-
-        <h5 className={css('title')}>Schema</h5>
-        <Json body={structure.schema} />
       </div>
     )
   }
@@ -83,7 +90,7 @@ export default class StructureViewer extends Base {
     const palette = defaultPalette
     return {
       wrap: {
-        marginBottom: 40
+        marginBottom: 30
       },
       title: {
         marginTop: 30
@@ -119,12 +126,12 @@ export default class StructureViewer extends Base {
   }
 }
 
-StructureViewer.propTypes = {
-  structure: PropTypes.object.isRequired,
+MetadataViewer.propTypes = {
+  metadata: PropTypes.object.isRequired,
   stdOnly: PropTypes.bool
 }
 
-StructureViewer.defaultProps = {
-  structure: {},
+MetadataViewer.defaultProps = {
+  metadata: {},
   stdOnly: false
 }
