+++ lib/components/dataset/Overview.js
@@ -1,21 +1,23 @@
 import React from 'react'
+import { NavLink } from 'react-router-dom'
 import PropTypes from 'prop-types'
-import MetaProps from '../propTypes/metaProps.js'
-import Base from './Base'
-import TagList from './TagList'
+import MetaProps from '../../propTypes/metaProps.js'
+import Base from '../Base'
+import TagList from '../TagList'
 import StatsLineLarge from './StatsLineLarge'
-import CommitLine from './CommitLine'
-import ComponentTable from './ComponentTable'
-import MetadataViewer from './MetadataViewer'
+import CommitItem from '../item/CommitItem'
+import Button from '../Button'
 
-import { defaultPalette } from '../propTypes/palette'
+import { defaultPalette } from '../../propTypes/palette'
 
-export default class DatasetSummary extends Base {
+export default class Overview extends Base {
   constructor (props) {
     super(props);
     [
       'filterFields',
-      'renderFields'
+      'renderFields',
+      'renderTitle',
+      'renderMeta'
     ].forEach((m) => { this[m] = this[m].bind(this) })
   }
 
@@ -114,39 +116,70 @@ export default class DatasetSummary extends Base {
       })
   }
 
+  renderTitle (css, title, link) {
+    const { url } = this.props
+    return (
+      <div className={css('flex')}>
+        <h5>{title}</h5><NavLink className={css('nav')} to={url + '/' + link}>View Full</NavLink>
+      </div>
+    )
+  }
+
+  renderMeta (css, meta, name) {
+    if (!meta) {
+      if (name) {
+        return (
+          <h3 className={css('name')}>{name}</h3>
+        )
+      }
+      return
+    }
+    return (
+      <div className={css('wrap')}>
+        <div className={css('titleWrap')}>
+          <h3 className={css('name')}>{ (meta && meta.title) || name}</h3>
+          {/* <div className={css('open')}> open data</div> */}
+        </div>
+        {meta && meta.keywords ? <TagList tags={meta.keywords} /> : undefined}
+        {meta && meta.description ? <p>{meta.description}</p> : undefined}
+        {meta && meta.downloadPath ? <a>{meta.downloadPath}</a> : undefined}
+      </div>
+    )
+  }
+
   template (css) {
-    const { datasetRef } = this.props
+    const { datasetRef, profile, url } = this.props
 
     if (datasetRef === undefined || datasetRef.dataset === undefined) {
-      return (<p>No summary found for this dataset.</p>)
+      return (<p>No overview found for this dataset.</p>)
     }
 
-    const { dataset } = datasetRef
+    const { dataset, name } = datasetRef
     const { meta, structure, commit } = dataset
 
     let md = meta
     if (typeof meta === 'string') {
-      md = { path: meta }
+      md = undefined
+    }
+
+    const index = url.indexOf('/overview')
+    var newUrl = url.slice(0, index)
+    if (index === -1) {
+      newUrl = url
     }
 
     const updated = commit ? commit.timestamp : undefined
 
     return (
-      <div className={css('flex')}>
-        <div className={css('meta')} >
-          <div className={css('titleWrap')}>
-            <h2 className={css('name')}>{ md ? md.title : datasetRef.name}</h2>
-            {/* <div className={css('open')}> open data</div> */}
-          </div>
-          {md && md.description && <p>{md.description}</p>}
-          {md && md.keywords && <TagList tags={md.keywords} />}
-          {md && md.downloadPath && <a>{md.downloadPath}</a>}
-        </div>
+      <div className={css('wrap')}>
+        {/* md && this.renderTitle(css, 'Meta', 'meta') */}
+        {this.renderMeta(css, md, name)}
+        {/* structure && updated && this.renderTitle(css, 'Structure', 'structure') */}
         {structure && updated && <StatsLineLarge stats={this.stats(structure)} updated={updated} />}
-        {commit && <CommitLine path={datasetRef.path} commit={commit} />}
+        {/* commit && this.renderTitle(css, 'Commit', 'commit') */}
+        {commit && <CommitItem data={datasetRef} profile={profile} /> }
         {/* this.renderFields(md, css) */}
-        <ComponentTable dataset={datasetRef.dataset} url={this.props.url} date={updated} />
-        <MetadataViewer metadata={md} />
+        <div className={css('buttonWrap')}><Button color='a' large full link={`${newUrl}/body`} text='TAKE ME TO THE DATA!' /></div>
       </div>
     )
   }
@@ -186,15 +219,27 @@ export default class DatasetSummary extends Base {
       },
       name: {
         margin: 0
+      },
+      flex: {
+        marginTop: 40,
+        display: 'flex',
+        alignItems: 'baseline'
+      },
+      nav: {
+        marginLeft: 5,
+        fontSize: 14
+      },
+      buttonWrap: {
+        marginTop: 40
       }
     }
   }
 }
 
-DatasetSummary.propTypes = {
+Overview.propTypes = {
   meta: MetaProps,
   onClickEdit: PropTypes.func
 }
 
-DatasetSummary.defaultProps = {
+Overview.defaultProps = {
 }
