+++ lib/components/dataset/Render.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Base from './Base'
+import Base from '../Base'
 
 export default class Render extends Base {
   template (css) {
