+++ lib/components/dataset/Structure.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import { structureProps } from '../propTypes/datasetRefProps.js'
-import Base from './Base'
+import { structureProps } from '../../propTypes/datasetRefProps.js'
+import Base from '../Base'
 // import Schema from './schema/schema'
 import StructureViewer from './StructureViewer'
