+++ lib/components/dataset/StructureViewer.js
@@ -1,34 +1,24 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import { defaultPalette } from '../propTypes/palette'
+import { defaultPalette } from '../../propTypes/palette'
 
-import Base from './Base'
-import Json from './Json'
+import Base from '../Base'
+import Json from '../Json'
 
-export default class MetadataViewer extends Base {
+export default class StructureViewer extends Base {
   constructor (props) {
     super(props)
 
     this.stdFields = [
-      'title',
-      'theme',
-      'keywords',
-      'description',
-      'license',
-      'accessPath',
-      'language',
-      'citations',
-      'contributors',
-      'accrualPeriodicity',
-      'downloadPath',
-      'homePath',
-      'identifier',
-      'readmePath',
-      'version'
+      'checksum',
+      'errCount',
+      'entries',
+      'format',
+      'length'
     ]
 
-    this.ignoreFields = ['qri', 'path'];
+    this.ignoreFields = ['qri', 'path', 'schema'];
 
     ['renderTable', 'renderValue'].forEach((m) => { this[m] = this[m].bind(this) })
   }
@@ -66,22 +56,25 @@ export default class MetadataViewer extends Base {
   }
 
   template (css) {
-    const { metadata, stdOnly } = this.props
+    const { structure, stdOnly } = this.props
 
-    const std = this.stdFields.filter((key) => !!metadata[key])
-    const extra = Object.keys(metadata).filter((key) => {
+    const std = this.stdFields.filter((key) => !!structure[key])
+    const extra = Object.keys(structure).filter((key) => {
       return !(~this.stdFields.findIndex((sKey) => (key === sKey)) || ~this.ignoreFields.findIndex((iKey) => (key === iKey)))
     })
 
     return (
       <div className={css('wrap')}>
-        <h5 className={css('title')}>Standard Metadata</h5>
-        {this.renderTable(css, std, metadata)}
+        <h5 className={css('title')}>Structure</h5>
+        {this.renderTable(css, std, structure)}
 
         {!stdOnly && (extra.length > 0) && <div>
-          <h5 className={css('title')}>Additional Metadata</h5>
-          {this.renderTable(css, extra, metadata)}
+          <h5 className={css('title')}>Schema</h5>
+          {this.renderTable(css, extra, structure)}
         </div>}
+
+        <h5 className={css('title')}>Schema</h5>
+        <Json body={structure.schema} />
       </div>
     )
   }
@@ -90,7 +83,7 @@ export default class MetadataViewer extends Base {
     const palette = defaultPalette
     return {
       wrap: {
-        marginBottom: 30
+        marginBottom: 40
       },
       title: {
         marginTop: 30
@@ -126,12 +119,12 @@ export default class MetadataViewer extends Base {
   }
 }
 
-MetadataViewer.propTypes = {
-  metadata: PropTypes.object.isRequired,
+StructureViewer.propTypes = {
+  structure: PropTypes.object.isRequired,
   stdOnly: PropTypes.bool
 }
 
-MetadataViewer.defaultProps = {
-  metadata: {},
+StructureViewer.defaultProps = {
+  structure: {},
   stdOnly: false
 }
