+++ lib/components/dataset/Viz.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import Base from './Base'
-import Spinner from './Spinner'
-import CodeViewer from './CodeViewer'
+import Base from '../Base'
+import Spinner from '../Spinner'
+import CodeViewer from '../CodeViewer'
 
 export default class Viz extends Base {
   componentWillMount () {
