+++ lib/containers/Commits.js
@@ -5,7 +5,7 @@ import { selectHistory } from '../selectors/history'
 import { selectIsFetching, selectPageCount, selectFetchedAll } from '../selectors/pagination'
 import { selectSessionProfileId } from '../selectors/session'
 
-import Commits from '../components/Commits'
+import Commits from '../components/dataset/Commits'
 
 const CommitsContainer = connect(
   (state, ownProps) => {
