+++ lib/containers/Dataset.js
@@ -11,7 +11,7 @@ import { selectIsTransfering } from '../selectors/transfers'
 import { showModal, hideModal } from '../actions/app'
 import { loadProfileByName } from '../actions/profiles'
 
-import Dataset from '../components/Dataset'
+import Dataset from '../components/dataset/Dataset'
 
 const DatasetContainer = connect(
   (state, ownProps) => {
